Updating CRUD Products endpoints. H4 - Activity 2 - NestJS Routing - M4

+++ ecommerce-marcos-anselmo-portillo/src/products/products.controller.ts
@@ -1,4 +1,13 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import {
+  Controller,
+  Get,
+  Post,
+  Body,
+  Param,
+  Delete,
+  NotFoundException,
+  Put,
+} from '@nestjs/common';
 import { ProductsService } from './products.service';
 import { CreateProductDto } from './dto/create-product.dto';
 import { UpdateProductDto } from './dto/update-product.dto';
@@ -9,7 +18,12 @@ export class ProductsController {
 
   @Post()
   create(@Body() createProductDto: CreateProductDto) {
-    return this.productsService.create(createProductDto);
+    const productId = this.productsService.create(createProductDto);
+    return {
+      statusCode: 201,
+      message: 'Product created successfully',
+      productId: productId,
+    };
   }
 
   @Get()
@@ -19,16 +33,38 @@ export class ProductsController {
 
   @Get(':id')
   findOne(@Param('id') id: string) {
-    return this.productsService.findOne(+id);
+    const product = this.productsService.findOne(+id);
+    if (!product) {
+      throw new NotFoundException(`Product with id ${id} not found`);
+    }
+    return product;
   }
 
-  @Patch(':id')
+  @Put(':id')
   update(@Param('id') id: string, @Body() updateProductDto: UpdateProductDto) {
-    return this.productsService.update(+id, updateProductDto);
+    let product = this.productsService.findOne(+id);
+    if (!product) {
+      throw new NotFoundException(`Product with id ${id} not found`);
+    }
+    this.productsService.update(+id, updateProductDto);
+    return {
+      statusCode: 200,
+      message: 'Product updated successfully',
+      productId: id,
+    };
   }
 
   @Delete(':id')
   remove(@Param('id') id: string) {
-    return this.productsService.remove(+id);
-  }
+    const product = this.productsService.findOne(+id);
+    if (!product) {
+      throw new NotFoundException(`Product with id ${id} not found`);
+    }
+    this.productsService.remove(+id);
+    return {
+      statusCode: 200,
+      message: 'Product deleted successfully',
+      productId: id,
+    }
+  } 
 }
